Updated the reducer to context
