+++ to-do-list/src/App.js
@@ -1,10 +1,13 @@
 import './App.css';
 import TaskList from './components/TaskList/TaskList';
+import { TasksProvider } from './hooks/TasksContext';
 
 function App() {
   return (
     <div className="App">
-      <TaskList/>
+      <TasksProvider>
+        <TaskList/>
+      </TasksProvider>
     </div>
   );
 }
