+++ to-do-list/src/components/Task/Task.jsx
@@ -2,15 +2,17 @@ import {React, useState} from 'react'
 import {ImCheckboxUnchecked, ImCheckboxChecked} from 'react-icons/im';
 import {AiFillDelete, AiFillEdit} from 'react-icons/ai';
 import "./Task.css"
-import { ACTIONS } from '../TaskList/TaskList';
+import { ACTIONS, useDispatch} from '../../hooks/TasksContext';
 import { inValidTask } from '../../Regex';
 
 
-export default function Task({task, dispatch}) {
+export default function Task({task}) {
     
     const[edit, setEdit] = useState(false);
     const[title, setTitle] = useState("");
-
+    
+    const dispatch = useDispatch();
+    
     function onKeyUp(e) {
         if (e.key === 'Enter' && !inValidTask.test(title)  ) {
             dispatch({type: ACTIONS.EDIT_TASK, payload:{id: task.id, title:title }})
